# 1.8x Faster Partial Caching Code